# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: uplook API
  version: 1.0.0
  summary: "「uplook」小程序 / 手机端 v1 接口契约"
  description: |
    uplook 的最终 /v1 传输契约。质量主链使用不可变 Report、PlanSet、
    RenderRun、Selection、Execution 与公开 Operation；客户端不读取内部 Task。
    生成媒体来源只由 source_kind 强类型确定。
  license:
    name: UNLICENSED
    identifier: LicenseRef-UNLICENSED
servers:
- url: https://{host}
  description: 线上/预发网关（host 由部署环境注入）
  variables:
    host:
      default: api.zhanshimian.internal
      description: 网关域名占位符，本地开发请使用下方 127.0.0.1 服务地址
- url: http://127.0.0.1:58000
  description: 本地开发（docker-compose 暴露的服务端口）
...
security:
- bearerAuth: []
paths:
...
  "/v1/auth/wechat":
    post:
      operationId: wechatLogin
      summary: 微信小程序 code 登录
      description: 用 wx.login 的 code 换取会话；code 无效返回 401 wechat_code_invalid，微信侧限频 429 wechat_login_limited，微信不可用 502 wechat_unavailable。公开端点。
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - code
              properties:
                code:
                  type: string
                  description: wx.login 获取的登录凭证
                nickname:
                  type: string
                  description: 首次登录的昵称（已注册用户忽略）
      responses:
        '201':
          description: 登录成功，返回会话令牌与用户
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/Session"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '502':
          "$ref": "#/components/responses/BadGateway"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/auth/dev":
    post:
      operationId: devLogin
      summary: 开发环境快捷登录
      description: 仅开发/预发环境开放；生产环境固定返回 404 not_found。公开端点。
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  description: 昵称（缺省自动生成）
      responses:
        '201':
          description: 登录成功，返回会话
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/Session"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '404':
          "$ref": "#/components/responses/NotFound"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/auth/sms/request":
    post:
      operationId: requestSmsCode
      summary: 请求短信验证码
      description: 向手机号发送 6 位验证码。限频：60s 冷却、单 IP 10 条/分钟、单号 5 条/小时，超限返回 429 rate_limited；验证码仅以摘要存储。公开端点。
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - phone
              properties:
                phone:
                  type: string
                  pattern: "^1[0-9]{10}$"
                  description: 中国大陆手机号
      responses:
        '202':
          description: 验证码已受理下发；cooldown_seconds 内重复请求将被限频
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/SmsRequestAck"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/auth/sms/verify":
    post:
      operationId: verifySmsCode
      summary: 短信验证码登录
      description: 校验验证码并登录/注册；手机号已绑定其他端（如 Apple 同号）时自动合并到同一账户，数据互通。验证码错误或过期返回 400 validation_error。公开端点。
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - phone
              - code
              properties:
                phone:
                  type: string
                  pattern: "^1[0-9]{10}$"
                  description: 中国大陆手机号
                code:
                  type: string
                  pattern: "^[0-9]{6}$"
                  description: 6 位短信验证码
      responses:
        '201':
          description: 登录成功（或合并至已有账户），返回会话
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/Session"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/auth/apple":
    post:
      operationId: appleLogin
      summary: iOS Apple 追加登录
      description: 以 Sign in with Apple 的 identityToken 登录；手机号与已有账户一致时自动合并（跨端数据互通）。identityToken 无效返回 401 unauthorized。公开端点。
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - identity_token
              properties:
                identity_token:
                  type: string
                  description: Sign in with Apple 返回的 identityToken
      responses:
        '201':
          description: 登录/合并成功，返回会话
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/Session"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/auth/wechat-app":
    post:
      operationId: wechatAppLogin
      summary: Android 微信开放平台追加登录
      description: Android 端微信 SDK code 登录；错误语义与小程序登录一致（wechat_code_invalid / wechat_login_limited / wechat_unavailable）。公开端点。
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - code
              properties:
                code:
                  type: string
                  description: 微信开放平台 OAuth code
                nickname:
                  type: string
                  description: 首次登录的昵称
      responses:
        '201':
          description: 登录成功，返回会话
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/Session"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '429':
          "$ref": "#/components/responses/TooManyRequests"
        '502':
          "$ref": "#/components/responses/BadGateway"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/auth/session":
    delete:
      operationId: logout
      summary: 退出登录
      description: 删除当前会话行（仅当前端，不影响其他端已登录会话）。幂等：会话已失效也返回 204。
      tags:
      - Auth
      responses:
        '204':
          description: 已退出，无响应体
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/me":
    get:
      operationId: getMe
      summary: 当前账户与多端身份
      description: 返回账户信息与 identities 多端身份列表（wechat_miniapp / wechat_app / apple / phone），按绑定时间排序。
      tags:
      - Me
      responses:
        '200':
          description: 账户与身份列表
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/MeAccount"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '500':
          "$ref": "#/components/responses/InternalError"
    patch:
      operationId: updateMe
      summary: 修改当前账户称呼或头像
      tags:
      - Me
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  minLength: 1
                  maxLength: 20
                avatar_media_id:
                  type: string
                  description: 已上传媒体 ID，用作展示头像，不替换建档照片
      responses:
        '200':
          description: 更新后的账户
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/MeAccount"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/me/profile":
    get:
      operationId: getMyProfile
      summary: 读取补充资料
      description: 返回身高/职业/预算与选填的体重、三围；从未填写时 data 为 null（客户端应引导进入资料页）。
      tags:
      - Me
      responses:
        '200':
          description: 补充资料；未填写时 data 为 null
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      oneOf:
                      - "$ref": "#/components/schemas/UserProfile"
                      - type: 'null'
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '500':
          "$ref": "#/components/responses/InternalError"
    put:
      operationId: updateMyProfile
      summary: 保存补充资料
      description: 全量保存（幂等，重复提交同一资料结果一致）；身高/职业/预算必填，体重与三围选填。
      tags:
      - Me
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UserProfile"
      responses:
        '200':
          description: 保存后的资料
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/UserProfile"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/me/data":
    delete:
      operationId: deleteMyData
      summary: 删除我的全部数据
      description: 注销清除：删除该用户全部业务数据与会话（含各端身份），立即生效且不可恢复，返回 204。
      tags:
      - Me
      responses:
        '204':
          description: 已删除，无响应体
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/home/bootstrap":
    get:
      operationId: getHomeBootstrap
      summary: 获取首页质量闭环聚合
      tags:
      - Home
      responses:
        '200':
          description: 首页聚合
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - data
                properties:
                  data:
                    "$ref": "#/components/schemas/HomeBootstrap"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '500':
          "$ref": "#/components/responses/InternalError"
      description: 首页稳定读模型；active_operations 只包含公开 OperationRef。
  "/v1/diagnostics":
    post:
      operationId: createDiagnosis
      summary: 穿搭诊断（同步）
      description: 同步诊断，201 立即返回结论（无异步任务）。kind=outfit 穿搭诊断（media kind=outfit），kind=purchase 购单品决策（media kind=product）。替代旧 RPC 风格 tools/run。
      tags:
      - Diagnostics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - kind
              - media_id
              properties:
                kind:
                  type: string
                  enum:
                  - outfit
                  - purchase
                  description: 诊断类型
                media_id:
                  type: string
                  description: 待诊断照片的媒体 ID
                scene:
                  type: string
                  description: 场景（选填）
                report_id:
                  type: string
                  description: 关联报告（选填，缺省用当前报告）
      responses:
        '201':
          description: 诊断结论（findings 带锚点、options 给可执行建议）
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/Diagnosis"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '422':
          description: 照片门禁拒识（error.code=photo_rejected，message 为可公开原因；不落库，客户端引导换一张）
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorEnvelope"
        '500':
          "$ref": "#/components/responses/InternalError"
  "/v1/diagnostics/latest":
    get:
      operationId: getLatestDiagnosis
      summary: 最近一条诊断
      description: 按 kind 返回当前用户最新一条诊断（outfit / purchase）；客户端离开同步诊断页后据此恢复结论。无记录时 404。
      tags:
      - Diagnostics
      parameters:
      - name: kind
        in: query
        required: true
        description: 诊断类型
        schema:
          type: string
          enum:
          - outfit
          - purchase
      responses:
        '200':
          description: 最近一条诊断
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/Diagnosis"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '404':
          "$ref": "#/components/responses/NotFound"
        '500':
          "$ref": "#/components/responses/InternalError"
...
  "/v1/daily/collection/stats":
    get:
      operationId: getDailyCollectionStats
      summary: 手册分格计数
      description: 全部分格（七格 + general）的条数，客户端徽标用它展示。
      tags:
      - Daily
      responses:
        '200':
          description: 分类计数
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/Envelope"
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      "$ref": "#/components/schemas/DailyCollectionStats"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '500':
          "$ref": "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session-token
      description: 'Authorization: Bearer <token>，token 来自登录端点返回的 Session.token'
  headers:
    XRequestId:
      description: 请求追踪 ID（客户端可透传，服务端缺省自动生成；与错误体 request_id 一致）
      schema:
        type: string
...
  responses:
    BadRequest:
      description: 参数或请求体校验失败（error.code=validation_error）
      headers:
        X-Request-ID:
          "$ref": "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Envelope:
      type: object
      description: 统一成功响应包装；异步创建在顶层返回 data 与 operation。
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          "$ref": "#/components/schemas/Error"
...
    Session:
      type: object
      required:
      - token
      - expires_at
      - user
      properties:
        token:
          type: string
          description: 会话令牌（Bearer）
        expires_at:
          type: string
          format: date-time
          description: 过期时间（滑动续期：余量 <7 天自动续 30 天）
        user:
          "$ref": "#/components/schemas/User"
...